Remove "Apis from other specs" section
In the respec-version, this section served as a place where links to
other specs would be <dfn>ed so that <a> links within this spec could
link here.

In the bikeshed version, we just link directly to the external specs
(so that readers don’t have to follow multiple links to get to the
definition).

Thus, there is no need for this section.