Fix algorithm for firing a clipboard event
This part of the algorithm got messed up during the transition to
bikeshed so both clauses were set to be the same.
A previous CL merged the 2 clauses (since they were identical).
This CL recovers how |target| was set in the original algorithm.
It also regenerated the HTML file for the previous few CLs.